requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
PyYAML==3.12
six==1.10.0
requests==2.18.4
selenium==4.7.2
PyMySQL==0.7.11
Appium-Python-Client==2.7.1
Pillow==8.1.2
xlsxwriter==1.0.4
requests-toolbelt==0.8.0
wrapcache==1.0.8
termcolor==2.1.1